LLDP: enable configuration on the fly
`make_lldp_manager` is implemented as a lambda and receives a list of host interfaces at creation of the handler.
If i.e. the IP address of an interface changes the lldp-manager has to destroyed and recreated. By re-reading the `host_interfaces` from the `model.settings` this can be avoided and `configure_interfaces` can be called instead.